Require user authentication to call chat endpoints
